Extract ContributionIndexTableModel as outer class
